requirements.txt
nltk==3.9.0
librosa==0.10.2
torch==2.7.0
torchvision==0.22.0
streamlit==1.38.0
numpy==2.0.0
pillow==10.4.0
scikit-learn==1.5.2
matplotlib==3.9.2
gensim==4.3.3
